+++ oap-server/server-storage-plugin/storage-banyandb-plugin/src/main/java/org/apache/skywalking/oap/server/storage/plugin/banyandb/BanyanDBZipkinQueryDAO.java
@@ -36,6 +36,7 @@ import org.apache.skywalking.banyandb.v1.client.RowEntity;
 import org.apache.skywalking.banyandb.v1.client.StreamQuery;
 import org.apache.skywalking.banyandb.v1.client.StreamQueryResponse;
 import org.apache.skywalking.banyandb.v1.client.TimestampRange;
+import org.apache.skywalking.oap.server.core.analysis.DownSampling;
 import org.apache.skywalking.oap.server.core.query.input.Duration;
 import org.apache.skywalking.oap.server.core.storage.query.IZipkinQueryDAO;
 import org.apache.skywalking.oap.server.core.zipkin.ZipkinServiceRelationTraffic;
@@ -84,8 +85,9 @@ public class BanyanDBZipkinQueryDAO extends AbstractBanyanDBDAO implements IZipk
 
     @Override
     public List<String> getServiceNames() throws IOException {
+        MetadataRegistry.Schema schema = MetadataRegistry.INSTANCE.findMetadata(ZipkinServiceTraffic.INDEX_NAME, DownSampling.Minute);
         MeasureQueryResponse resp =
-            query(ZipkinServiceTraffic.INDEX_NAME,
+            query(schema,
                   SERVICE_TRAFFIC_TAGS,
                   Collections.emptySet(), new QueryBuilder<MeasureQuery>() {
 
@@ -104,8 +106,9 @@ public class BanyanDBZipkinQueryDAO extends AbstractBanyanDBDAO implements IZipk
 
     @Override
     public List<String> getRemoteServiceNames(final String serviceName) throws IOException {
+        MetadataRegistry.Schema schema = MetadataRegistry.INSTANCE.findMetadata(ZipkinServiceRelationTraffic.INDEX_NAME, DownSampling.Minute);
         MeasureQueryResponse resp =
-            query(ZipkinServiceRelationTraffic.INDEX_NAME,
+            query(schema,
                   REMOTE_SERVICE_TRAFFIC_TAGS,
                   Collections.emptySet(), new QueryBuilder<MeasureQuery>() {
 
@@ -127,8 +130,9 @@ public class BanyanDBZipkinQueryDAO extends AbstractBanyanDBDAO implements IZipk
 
     @Override
     public List<String> getSpanNames(final String serviceName) throws IOException {
+        MetadataRegistry.Schema schema = MetadataRegistry.INSTANCE.findMetadata(ZipkinServiceSpanTraffic.INDEX_NAME, DownSampling.Minute);
         MeasureQueryResponse resp =
-            query(ZipkinServiceSpanTraffic.INDEX_NAME,
+            query(schema,
                   SPAN_TRAFFIC_TAGS,
                   Collections.emptySet(), new QueryBuilder<MeasureQuery>() {